backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
spacy==3.7.2
openai==1.3.5
python-dotenv==1.0.0
pydantic==2.5.0
